Add constructor for msg content, move Deserialize to content_serde.rs
