+++ ruma-events/src/room/message.rs
@@ -7,16 +7,11 @@ use ruma_events_macros::MessageEventContent;
 #[cfg(feature = "unstable-pre-spec")]
 use ruma_identifiers::{DeviceIdBox, UserId};
 use ruma_serde::StringEnum;
-use serde::{
-    de::{self, Error},
-    Deserialize, Serialize,
-};
-use serde_json::{value::RawValue as RawJsonValue, Value as JsonValue};
+use serde::{Deserialize, Serialize};
+use serde_json::Value as JsonValue;
 
 #[cfg(feature = "unstable-pre-spec")]
 use crate::key::verification::VerificationMethod;
-
-use crate::from_raw_json_value;
 
 #[cfg(feature = "unstable-pre-spec")]
 use super::relationships::{Annotation, Reference, RelationJsonRepr, Replacement};
@@ -60,6 +55,17 @@ pub struct MessageEventContent {
 }
 
 impl MessageEventContent {
+    /// A convenience constructor to create a `MessageEventContent` with the given
+    /// `MessageType`.
+    pub fn new(msgtype: MessageType) -> Self {
+        Self {
+            msgtype,
+            relates_to: None,
+            #[cfg(feature = "unstable-pre-spec")]
+            new_content: None,
+        }
+    }
+
     /// A convenience constructor to create a plain text message.
     pub fn text_plain(body: impl Into<String>) -> Self {
         Self {
@@ -80,7 +86,7 @@ impl MessageEventContent {
         }
     }
 
-    /// A convenience constructor to create an plain text notice.
+    /// A convenience constructor to create a plain text notice.
     pub fn notice_plain(body: impl Into<String>) -> Self {
         Self {
             msgtype: MessageType::Notice(NoticeMessageEventContent::plain(body)),
@@ -102,7 +108,7 @@ impl MessageEventContent {
 }
 
 /// The content that is specific to each message type variant.
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Serialize)]
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
 #[serde(tag = "msgtype")]
 pub enum MessageType {
@@ -140,6 +146,17 @@ pub enum MessageType {
     /// A custom message.
     #[doc(hidden)]
     _Custom(CustomEventContent),
+}
+
+impl From<MessageType> for MessageEventContent {
+    fn from(msgtype: MessageType) -> Self {
+        Self {
+            msgtype,
+            relates_to: None,
+            #[cfg(feature = "unstable-pre-spec")]
+            new_content: None,
+        }
+    }
 }
 
 /// Enum modeling the different ways relationships can be expressed in a
@@ -615,96 +632,4 @@ pub struct CustomEventContent {
     /// Remaining event content
     #[serde(flatten)]
     pub data: BTreeMap<String, JsonValue>,
-}
-
-#[doc(hidden)]
-#[derive(Debug, Deserialize)]
-pub struct MessageDeHelper {
-    /// the Matrix message type string "m.whatever".
-    pub msgtype: String,
-
-    #[serde(rename = "m.relates_to")]
-    relates_to: Option<Box<RawJsonValue>>,
-
-    #[cfg(feature = "unstable-pre-spec")]
-    #[serde(rename = "m.new_content")]
-    new_content: Option<Box<RawJsonValue>>,
-}
-
-impl<'de> de::Deserialize<'de> for MessageEventContent {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: de::Deserializer<'de>,
-    {
-        let json = Box::<RawJsonValue>::deserialize(deserializer)?;
-        #[cfg(feature = "unstable-pre-spec")]
-        let MessageDeHelper { msgtype, relates_to, new_content } = from_raw_json_value(&json)?;
-        #[cfg(not(feature = "unstable-pre-spec"))]
-        let MessageDeHelper { msgtype, relates_to } = from_raw_json_value(&json)?;
-
-        Ok(match msgtype.as_str() {
-            "m.audio" => Self {
-                msgtype: MessageType::Audio(from_raw_json_value(&json)?),
-                relates_to: relates_to.map(|json| from_raw_json_value(&json)).transpose()?,
-                #[cfg(feature = "unstable-pre-spec")]
-                new_content: new_content.map(|json| from_raw_json_value(&json)).transpose()?,
-            },
-            "m.emote" => Self {
-                msgtype: MessageType::Emote(from_raw_json_value(&json)?),
-                relates_to: relates_to.map(|json| from_raw_json_value(&json)).transpose()?,
-                #[cfg(feature = "unstable-pre-spec")]
-                new_content: new_content.map(|json| from_raw_json_value(&json)).transpose()?,
-            },
-            "m.file" => Self {
-                msgtype: MessageType::File(from_raw_json_value(&json)?),
-                relates_to: relates_to.map(|json| from_raw_json_value(&json)).transpose()?,
-                #[cfg(feature = "unstable-pre-spec")]
-                new_content: new_content.map(|json| from_raw_json_value(&json)).transpose()?,
-            },
-            "m.image" => Self {
-                msgtype: MessageType::Image(from_raw_json_value(&json)?),
-                relates_to: relates_to.map(|json| from_raw_json_value(&json)).transpose()?,
-                #[cfg(feature = "unstable-pre-spec")]
-                new_content: new_content.map(|json| from_raw_json_value(&json)).transpose()?,
-            },
-            "m.location" => Self {
-                msgtype: MessageType::Location(from_raw_json_value(&json)?),
-                relates_to: relates_to.map(|json| from_raw_json_value(&json)).transpose()?,
-                #[cfg(feature = "unstable-pre-spec")]
-                new_content: new_content.map(|json| from_raw_json_value(&json)).transpose()?,
-            },
-            "m.notice" => Self {
-                msgtype: MessageType::Notice(from_raw_json_value(&json)?),
-                relates_to: relates_to.map(|json| from_raw_json_value(&json)).transpose()?,
-                #[cfg(feature = "unstable-pre-spec")]
-                new_content: new_content.map(|json| from_raw_json_value(&json)).transpose()?,
-            },
-            "m.server_notice" => Self {
-                msgtype: MessageType::ServerNotice(from_raw_json_value(&json)?),
-                relates_to: relates_to.map(|json| from_raw_json_value(&json)).transpose()?,
-                #[cfg(feature = "unstable-pre-spec")]
-                new_content: new_content.map(|json| from_raw_json_value(&json)).transpose()?,
-            },
-            "m.text" => Self {
-                msgtype: MessageType::Text(from_raw_json_value(&json)?),
-                relates_to: relates_to.map(|json| from_raw_json_value(&json)).transpose()?,
-                #[cfg(feature = "unstable-pre-spec")]
-                new_content: new_content.map(|json| from_raw_json_value(&json)).transpose()?,
-            },
-            "m.video" => Self {
-                msgtype: MessageType::Video(from_raw_json_value(&json)?),
-                relates_to: relates_to.map(|json| from_raw_json_value(&json)).transpose()?,
-                #[cfg(feature = "unstable-pre-spec")]
-                new_content: new_content.map(|json| from_raw_json_value(&json)).transpose()?,
-            },
-            #[cfg(feature = "unstable-pre-spec")]
-            "m.key.verification.request" => Self {
-                msgtype: MessageType::VerificationRequest(from_raw_json_value(&json)?),
-                relates_to: relates_to.map(|json| from_raw_json_value(&json)).transpose()?,
-                #[cfg(feature = "unstable-pre-spec")]
-                new_content: new_content.map(|json| from_raw_json_value(&json)).transpose()?,
-            },
-            _ => return Err(D::Error::custom("Invalid msgtype found")),
-        })
-    }
 }
